fix(media-area): fixing tests + code cleanup
