+++ bigbluebutton-html5/imports/ui/components/actions-bar/media-area/media-sharing/external-video/component.tsx
@@ -33,6 +33,8 @@ const intlMessages = defineMessages({
 const YOUTUBE_SHORTS_REGEX = new RegExp(/^(?:https?:\/\/)?(?:www\.)?(youtube\.com\/shorts)\/.+$/);
 const PANOPTO_MATCH_URL = /https?:\/\/([^/]+\/Panopto)(\/Pages\/Viewer\.aspx\?id=)([-a-zA-Z0-9]+)/;
 
+const YOUTUBE_REGEX = new RegExp(/^(?:https?:\/\/)?(?:www\.)?(youtube\.com|youtu.be)\/.+$/);
+
 interface ExternalVideoViewProps {
   intl: IntlShape;
   isSharingVideo: boolean;
@@ -59,6 +61,13 @@ const ExternalVideoView: React.FC<ExternalVideoViewProps> = ({
       if (m && m.length >= 4) {
         externalVideoUrl = `https://${m[1]}/Podcast/Social/${m[3]}.mp4`;
       }
+    }
+
+    if (YOUTUBE_REGEX.test(externalVideoUrl)) {
+      const YTUrl = new URL(externalVideoUrl);
+      YTUrl.searchParams.delete('list');
+      YTUrl.searchParams.delete('index');
+      externalVideoUrl = YTUrl.toString();
     }
 
     startExternalVideo({ variables: { externalVideoUrl } });
