+++ bigbluebutton-html5/imports/ui/components/app/component.jsx
@@ -18,7 +18,6 @@ import BannerBarContainer from '/imports/ui/components/banner-bar/container';
 import RaiseHandNotifier from '/imports/ui/components/raisehand-notifier/container';
 import ManyWebcamsNotifier from '/imports/ui/components/video-provider/many-users-notify/container';
 import AudioCaptionsSpeechContainer from '/imports/ui/components/audio/audio-graphql/audio-captions/speech/component';
-import UploaderContainer from '/imports/ui/components/presentation/presentation-uploader/container';
 import ScreenReaderAlertContainer from '../screenreader-alert/container';
 import ScreenReaderAlertAdapter from '../screenreader-alert/adapter';
 import WebcamContainer from '../webcam/component';
@@ -383,7 +382,7 @@ class App extends Component {
             <WebcamContainer />
             {
               !isNonMediaLayout
-                && <ExternalVideoPlayerContainer />
+              && <ExternalVideoPlayerContainer />
             }
             <GenericContentMainAreaContainer
               genericMainContentId={genericMainContentId}
@@ -415,7 +414,6 @@ class App extends Component {
             {(
               !hideNotificationToasts
               && isNotificationEnabled) && <PresentationUploaderToastContainer intl={intl} />}
-            <UploaderContainer />
             <BreakoutJoinConfirmationContainerGraphQL />
             <BBBLiveKitRoomContainer />
             <AudioContainer {...{
