--- bigbluebutton-html5/imports/ui/components/presentation/presentation-uploader/mutation.ts
@@ -1,15 +1,0 @@
-import { gql } from '@apollo/client';
-
-export const requestPresentationUploadTokenMutation = gql`
-  mutation RequestPresentationUploadToken($podId: String!, $filename: String!, $uploadTemporaryId: String!) {
-    presentationRequestUploadToken(
-      podId: $podId,
-      filename: $filename,
-      uploadTemporaryId: $uploadTemporaryId,
-    )
-  }
-`;
-
-export default {
-  requestPresentationUploadTokenMutation,
-};
