--- bigbluebutton-html5/imports/ui/components/presentation/presentation-uploader/presentation-download-dropdown/presentation-download-dropdown-wrapper/component.jsx
@@ -1,18 +1,0 @@
-import React from 'react';
-import PropTypes from 'prop-types';
-import Styled from './styles';
-
-const propTypes = {
-  children: PropTypes.shape({}).isRequired,
-};
-function PresentationDownloadDropdownWrapper({ children }) {
-  return (
-    <Styled.DropdownMenuWrapper>
-      {children}
-    </Styled.DropdownMenuWrapper>
-  );
-}
-
-PresentationDownloadDropdownWrapper.propTypes = propTypes;
-
-export default PresentationDownloadDropdownWrapper;
