--- bigbluebutton-html5/imports/ui/components/presentation/presentation-uploader/presentation-download-dropdown/presentation-download-dropdown-wrapper/styles.js
@@ -1,9 +1,0 @@
-import styled from 'styled-components';
-
-const DropdownMenuWrapper = styled.div`
-  display: inline-block;
-`;
-
-export default {
-  DropdownMenuWrapper,
-};
