--- bigbluebutton-html5/imports/ui/components/presentation/presentation-uploader/queries.ts
@@ -1,15 +1,0 @@
-import { gql } from '@apollo/client';
-
-export const getPresentationUploadToken = gql`
-  query getPresentationUploadToken ($uploadTemporaryId: String!){
-    pres_presentation_uploadToken(where: {uploadTemporaryId: {_eq: $uploadTemporaryId}}) {
-      presentationId
-      uploadTemporaryId
-      uploadToken
-    }
-  }
-`;
-
-export default {
-  getPresentationUploadToken,
-};
