+++ bigbluebutton-html5/public/locales/pt_BR.json
@@ -249,7 +249,7 @@
     "app.screenshare.screenshareRetryOtherEnvError": "Código {errorCode}. Não foi possível compartilhar a tela. Tente novamente usando um navegador ou dispositivo diferente.",
     "app.screenshare.screenshareUnsupportedEnv": "Código {errorCode}. O navegador não é compatível. Tente novamente usando um navegador ou dispositivo diferente.",
     "app.screenshare.screensharePermissionError": "Código {errorCode}. A permissão de compartilhamento de tela deve ser concedida.",
-    "app.screenshare.screenshareToastHelpLabel": "Saber mais", 
+    "app.screenshare.screenshareToastHelpLabel": "Saber mais",
     "app.cameraAsContent.presenterLoadingLabel": "Sua câmera está carregando",
     "app.cameraAsContent.viewerLoadingLabel": "A câmera do apresentador está carregando",
     "app.cameraAsContent.presenterSharingLabel": "Sua câmera está sendo transmitida",
@@ -693,11 +693,17 @@
     "app.switch.onLabel": "Ligado",
     "app.switch.offLabel": "Desligado",
     "app.talkingIndicator.ariaMuteDesc": "Selecione para silenciar o usuário",
-    "app.talkingIndicator.isTalking": "{userName} está falando",
-    "app.talkingIndicator.moreThanMaxIndicatorsTalking": "{userCount}+ estão falando",
-    "app.talkingIndicator.moreThanMaxIndicatorsWereTalking": "{userCount}+ estavam falando",
-    "app.talkingIndicator.wasTalking": "{userName} parou de falar",
-    "app.actionsBar.actionsDropdown.actionsLabel": "Ações",
+    "app.talkingIndicator.isTalking": "{0} está falando",
+    "app.talkingIndicator.moreThanMaxIndicatorsTalking": "{0}+ estão falando",
+    "app.talkingIndicator.moreThanMaxIndicatorsWereTalking": "{0}+ estavam falando",
+    "app.talkingIndicator.wasTalking": "{0} parou de falar",
+    "app.mediaSharing.modal.title": "Compartilhamento de mídia",
+    "app.mediaSharing.modal.slides": "Slides",
+    "app.mediaSharing.modal.videoLink": "Vídeo externo",
+    "app.mediaSharing.modal.mustBePresenter": "Você deve ser o apresentador para compartilhar conteúdo.",
+    "app.mediaSharing.modal.share": "Compartilhar",
+    "app.mediaSharing.modal.stopSharing": "Parar compartilhamento",
+    "app.actionsBar.actionsDropdown.actionsLabel": "Área de Mídia",
     "app.actionsBar.actionsDropdown.activateTimerStopwatchLabel": "Ativar temporizador/cronômetro",
     "app.actionsBar.actionsDropdown.deactivateTimerStopwatchLabel": "Desativar temporizador/cronômetro",
     "app.actionsBar.actionsDropdown.presentationLabel": "Gerenciar apresentações",
