+++ bigbluebutton-tests/playwright/breakout/join.js
@@ -67,7 +67,6 @@ class Join extends Create {
     await breakoutModPage.bringToFront();
 
     await breakoutModPage.closeAudioModal();
-    
     await breakoutModPage.waitForSelector(e.presentationTitle);
     return breakoutModPage;
   }
@@ -98,7 +97,7 @@ class Join extends Create {
     await this.modPage.waitAndClick(e.breakoutOptionsMenu);
     await this.modPage.waitAndClick(e.openUpdateBreakoutUsersModal);
     await this.modPage.dragDropSelector(e.attendeeNotAssigned, e.breakoutBox1);
-    await this.modPage.hasText(e.breakoutBox1, /Attendee/,  'should have the attendee name on the second breakout room box.');
+    await this.modPage.hasText(e.breakoutBox1, /Attendee/, 'should have the attendee name on the second breakout room box.');
     await this.modPage.waitAndClick(e.updateBreakoutRoomsButton);
 
     await this.userPage.hasElement(e.modalConfirmButton, 'should display the modal confirm button for the attendee to join the meeting');
@@ -169,18 +168,13 @@ class Join extends Create {
     // check if the notes were exported
     await this.modPage.hasElement(e.presentationUploadProgressToast, 'should display the presentation upload progress toast');
     await this.modPage.waitAndClick(e.mediaAreaButton);
-    const shareNotesPDF = await this.modPage.getLocatorByIndex(e.mediaAreaItem, 1);
+    await this.modPage.waitAndClick(e.managePresentations);
+    const shareNotesPDF = await this.modPage.getLocatorByIndex(e.presentationNames, 1);
     await expect(shareNotesPDF, 'should have the Notes name on the share notes pdf').toHaveText(/Notes/, { timeout: 30000 });
-    await expect(this.modPage.getLocatorByIndex(e.mediaAreaItem, 2)).toHaveText("Upload/Manage presentations"); //This checks if no other content was exported.
-    const expectedActionItems = [
-      "Default presentation",
-      "Exported breakout notes",
-      "Upload/Manage presentations",
-      "Share an external video",
-      "Share camera as content",
-    ];
-    await this.modPage.hasElementCount(e.mediaAreaItem, expectedActionItems.length);
-    await shareNotesPDF.click();
+    await this.modPage.hasElementCount(e.presentationNames, 2); //This checks if no other content was exported.
+    const shareNotesPDFThumbnail = await this.modPage.getLocatorByIndex(e.presentationThumbnails, 1);
+    await shareNotesPDFThumbnail.click();
+    await this.modPage.waitAndClick(e.sharePresentationButton);
     await hasCurrentPresentationToastElement(this.modPage, 'should display the current presentation toast when changing to the whiteboard exported file');
     // visual assertion
     await sleep(2000); // ensure whiteboard zoom is stabilized
@@ -219,18 +213,11 @@ class Join extends Create {
 
     await this.modPage.hasElement(e.presentationUploadProgressToast, 'should display the presentation upload progress toast', ELEMENT_WAIT_LONGER_TIME);
     await this.modPage.waitAndClick(e.mediaAreaButton);
-    const whiteboardPDF = await this.modPage.getLocatorByIndex(e.mediaAreaItem, 1);
+    await this.modPage.waitAndClick(e.managePresentations);
+    const whiteboardPDF = await this.modPage.getLocatorByIndex(e.presentationNames, 1);
     await expect(whiteboardPDF).toHaveText(/Whiteboard/, { timeout: 30000 });
-    await expect(this.modPage.getLocatorByIndex(e.mediaAreaItem, 2)).toHaveText("Upload/Manage presentations"); //This checks if no other content was exported.
-    const expectedActionItems = [
-      "Default presentation",
-      "Exported breakout notes",
-      "Upload/Manage presentations",
-      "Share an external video",
-      "Share camera as content",
-    ];
-    await this.modPage.hasElementCount(e.mediaAreaItem, expectedActionItems.length);
-    await this.modPage.press('Escape'); // close the actions menu
+    await this.modPage.hasElementCount(e.presentationNames, 2); //This checks if no other content was exported.
+    await this.modPage.press('Escape'); // close the media sharing menu
     await this.modPage.hasElement(e.presentationUploadProgressToast, 'should display the presentation upload progress toast with the exported whiteboard');
     await this.modPage.getLocator(e.presentationUploadProgressToast).click({
       position: {
@@ -240,7 +227,11 @@ class Join extends Create {
     });
     await this.modPage.wasRemoved(e.presentationUploadProgressToast, 'should have removed the presentation upload progress toast after clicking on it');
     await this.modPage.waitAndClick(e.mediaAreaButton);
-    await whiteboardPDF.click();
+    await this.modPage.waitAndClick(e.managePresentations);
+    const whiteboardPDFThumbnail = await this.modPage.getLocatorByIndex(e.presentationThumbnails, 1);
+    await whiteboardPDFThumbnail.click();
+    await this.modPage.waitAndClick(e.sharePresentationButton);
+    await this.modPage.press('Escape'); // close the media sharing menu
     await hasCurrentPresentationToastElement(this.modPage, 'should display the current presentation toast when changing to the whiteboard exported file');
     // visual assertion
     await sleep(2000); // ensure whiteboard zoom is stabilized
@@ -257,12 +248,12 @@ class Join extends Create {
     await this.userPage.hasElementEnabled(e.modalConfirmButton);
     await this.userPage.hasHiddenElementCount(e.roomOption, 2);
 
-    await this.userPage.getLocator(e.selectBreakoutRoomBtn).selectOption({index: 1});
+    await this.userPage.getLocator(e.selectBreakoutRoomBtn).selectOption({ index: 1 });
     await this.userPage.waitAndClick(e.modalConfirmButton);
 
     const breakoutUserPage = await this.userPage.getLastTargetPage(this.context);
     await breakoutUserPage.bringToFront();
-    await breakoutUserPage.hasElement(e.presentationTitle, 'should display the presentation title on the breakout room', ELEMENT_WAIT_LONGER_TIME);    
+    await breakoutUserPage.hasElement(e.presentationTitle, 'should display the presentation title on the breakout room', ELEMENT_WAIT_LONGER_TIME);
   }
 
   async breakoutWithDifferentPresentations() {
