+++ bigbluebutton-tests/playwright/presentation/util.js
@@ -32,20 +32,26 @@ async function uploadSinglePresentation(testPage, fileName, uploadTimeout = UPLO
     ?.style
     .backgroundImage
     .split('"')[1],
-  [e.currentSlideImg]);
+    [e.currentSlideImg]);
   await testPage.waitAndClick(e.mediaAreaButton);
   await testPage.waitAndClick(e.managePresentations);
   await testPage.hasElement(e.presentationFileUpload, 'should display the presentation space for uploading a new file, when the manage presentations is opened');
 
   await testPage.page.setInputFiles(e.presentationFileUpload, path.join(__dirname, `../core/media/${fileName}`));
-  await testPage.hasText('body', e.statingUploadPresentationToast, 'should display the toast message uploading the presentation');
 
-  await testPage.waitAndClick(e.confirmManagePresentation);
   await testPage.hasElement(e.presentationUploadProgressToast, 'should display the toast presentation upload progress after confirming the presentation to be uploaded');
+
+  await testPage.waitUntilHaveCountSelector(e.presentationThumbnails, 2, uploadTimeout) // wait for the upload to finish
+
+  // select and share the uploaded presentation
+  const newPDFThumbnail = await testPage.getLocatorByIndex(e.presentationThumbnails, 1);
+  await newPDFThumbnail.click();
+  await testPage.waitAndClick(e.sharePresentationButton);
+  await testPage.press('Escape'); // close the media sharing menu
   // ensures the current slide (uploaded file) is different from the previous slide - successful upload
   await testPage.page.waitForFunction(([selector, firstSlideSrc]) => {
     const currentSrc = document.querySelector(selector)
-    ?.style?.backgroundImage?.split('"')[1];
+      ?.style?.backgroundImage?.split('"')[1];
     return currentSrc != firstSlideSrc;
   }, [e.currentSlideImg, firstSlideSrc], {
     timeout: uploadTimeout,
@@ -59,12 +65,14 @@ async function uploadMultiplePresentations(testPage, fileNames, uploadTimeout = 
   await testPage.hasElement(e.presentationFileUpload, 'should display the modal for uploading a new presentation after opening the manage presentations');
 
   await testPage.page.setInputFiles(e.presentationFileUpload, fileNames.map((fileName) => path.join(__dirname, `../core/media/${fileName}`)));
-  await testPage.hasText('body', e.statingUploadPresentationToast, 'should display the toast of a presentation to be uploaded after selecting the files to upload');
-
-  await testPage.waitAndClick(e.confirmManagePresentation);
   await testPage.hasElement(e.presentationUploadProgressToast, 'should display a toast presentation upload progress after confirming the presentation to be uploaded');
-  await testPage.hasNElements(e.processingPresentationItem, fileNames.length, 'should display the presentation status info element with converting label after confirmation to upload the new file');
   await testPage.hasNElements(e.uploadDoneIcon, fileNames.length, 'should display the upload done icon after all presentations are successfully uploaded');
+  await testPage.waitUntilHaveCountSelector(e.presentationThumbnails, fileNames.length + 1, uploadTimeout) // wait for the uploads to finish
+  // select and share a new uploaded presentation
+  const newPDFThumbnail = await testPage.getLocatorByIndex(e.presentationThumbnails, 1);
+  await newPDFThumbnail.click();
+  await testPage.waitAndClick(e.sharePresentationButton);
+  await testPage.press('Escape'); // close the media sharing menu
   await hasCurrentPresentationToastElement(testPage, { timeout: uploadTimeout });
 }
